simulator/requirements.txt
kafka-python
numpy
scipy
pandas